Move all the startup printing stuff into the startup printing function
