+++ src/sandbox/print-status.js
@@ -1,11 +1,13 @@
-let { sep } = require('path')
+let { join, sep } = require('path')
+let { existsSync } = require('fs')
 let chalk = require('chalk')
+let { chars } = require('@architect/utils')
 let httpConfig = require('../http/_config')
 
 /**
- * Pretty print @http + @ws routes
+ * Pretty print all the things
  */
-module.exports = function prettyPrint (params, callback) {
+module.exports = function prettyPrint (params, start, callback) {
   let { cwd, inventory, ports, restart, update } = params
   let { http, ws } = inventory.inv
 
@@ -13,6 +15,9 @@ module.exports = function prettyPrint (params, callback) {
     return callback()
   }
 
+  /**
+   * All the HTTP/WS things
+   */
   let { apiType } = httpConfig(params)
   let padL = str => str.padStart(7)
   let padR = str => str.padEnd(45, '.')
@@ -51,5 +56,27 @@ module.exports = function prettyPrint (params, callback) {
     let link = chalk.green.bold.underline(`http://localhost:${ports.http}\n`)
     update.raw(`\n    ${link}`)
   }
+
+  /**
+   * Startup time
+   */
+  let finish = Date.now()
+  update.done(`Started in ${finish - start}ms`)
+  let isWin = process.platform.startsWith('win')
+  let ready = isWin
+    ? chars.done
+    : chalk.green.dim('❤︎')
+  let readyMsg = chalk.white('Local environment ready!')
+  update.raw(`${ready} ${readyMsg}\n`)
+
+  // Check aws-sdk installation status if installed globally
+  let dir = __dirname
+  if (!dir.startsWith(cwd) && !process.pkg) {
+    let awsDir = join(dir.split('@architect')[0], 'aws-sdk', 'package.json')
+    if (!existsSync(awsDir)) {
+      update.warn(`Possible global install of Architect without a global install of AWS-SDK, please run: npm i -g aws-sdk`)
+    }
+  }
+
   callback()
 }
